backend/requirements.txt
Flask
Flask-CORS
torch
torchvision
timm
opencv-python
Pillow
numpy
python-multipart
werkzeug